[DetailContent Get Data][Be][SP1][FT_03]: load data for table ALT_DETAIL_CONTENT

See merge request Vidal/com-intro-website-be!9

+++ src/main/java/com/altek/intro/dto/PageContentDTO.java
@@ -8,6 +8,6 @@ import javax.persistence.Column;
 public class PageContentDTO extends AbstractDTO{
     private String pageTitle;
     private String shortDescription;
-    private int menuId;
+    private Long menuId;
     private String type;
 }
