+++ src/main/java/com/altek/intro/repository/PageContentRepository.java
@@ -10,6 +10,6 @@ import java.util.List;
 public interface PageContentRepository extends AbstractRepository<PageContentEntity, Long>{
     @Query(value = "SELECT * FROM ALT_PAGE_CONTENT where ACTIVE_FLAG = 1", nativeQuery = true)
     List<PageContentEntity> findAll();
-    List<PageContentEntity> findAllByMenuId(int menuId);
+    List<PageContentEntity> findAllByMenuId(Long menuId);
 
 }
